Add PruneAggregationColumns Rule
Migrate PruneUnreferencedOutputs handling of AggregationNode to the
iterative optimizer.  This rule finds a AggregationNode that's under a
project, such that the project does not use all of the produced
aggregates, and removes unneeded aggregates.

Subsequently, PruneAggregationSourceColumns may notice that the child of
the AggregationNode provides columns that the AggregationNode does not
need, and insert a project between them which discards the unused
columns.