squash! Add spill for join
Replace commit message with:
---------------------------

Add spill for join

This adds spilling capability for inner and left outer join.
Spilling is initiated by memory revoking.

When spill is enabled, `HashBuilderOperator` (HBO) allocates revocable
memory.  Upon revoke memory request, the  input received so far is
spilled and any further input is spilled too.  Once all input is
received and spill has not occurred, a partition of the `LookupSource` is
build.  If the input has been spilled, the partition is not build, it's
marked in `PartitionedLookupSourceFactory` (PLSF) as spilled instead.

As it used to be, once all partitions of `LookupSource` are collected,
probe side is unblocked.  When `LookupJoinOperator` (LJO) receives
input, it checks whether spills has occurred for any of the partitions.
If it did, probe input page's positions that map to spilled partitions
are spilled (separately for each partition), the remaining ones continue
to the usual in-memory joining.  Once all LJOs finish input, unspilling
stage begins, wherein each `LookupSource`'s partition is in turn
unspilled, and corresponding rows of input are unspilled too (each LJO
unspills what it spilled previously) and joined in-memory, as usual.
This time, however, `LookupSource` is not partitioned, since there is
always single partition in memory at any given time.

Note, however, that spilling introduces another twist to the probing
stage.  HBO's memory may get revoked at any time.  If the memory is
revoked during probing stage, HBO spills to disk whole input it received
(it is kept by `PagesIndex`) and marks the partition as spilled in PLSF.
Marking is coordinated with LJOs so that removal of the partition from
the memory does not happen while it is used (while there is ongoing call
to `LookupJoinOperator.getOutput()`).  Technically, this is achieved by
the use of `ReadWriteLock` in PLSF.

Cautious reader will notice that a partition can be removed between the
calls `LookupJoinOperator.getOutput()` even though those calls can
operate one one input page, or even one input page position (if there
are multiple matching rows on the build side for single row on the probe
side).  Thus, it may happen that row being currently processed by
in-memory joining maps to a partition that is no longer in the memory.
In such a case, the row is stashed away until the unspilling stage,
together with information how far it was advanced in the collection of
the matching rows on the build side, so that joining can be resumed from
this stage.

All is neat and simple. ♠︎