Factor out ProjectOffPushDownRule
Add a generic ProjectOffPushDownRule to factor out the boilerplate of
rules migrated from PruneUnreferencedOutputs.

ProjectOffPushDownRule looks for a Project parent over a child of some
type N, such that the parent doesn't use all the output columns of the
child.  Given that situation, it invokes the abstract protected
pushDownProjectOff to possibly rewrite the child to produce fewer
outputs.