Factor out Util.pushDownProjectOff
Add a helper function Util.pushDownProjectOff to factor out the
boilerplate of rules migrated from PruneUnreferencedOutputs.

Util.pushDownProjectOff looks for a Project parent over a child of some
type N, such that the parent doesn't use all the output columns of the
child.  Given that situation, it invokes the provided BiFunction to
possibly rewrite the child to produce fewer outputs.