Replace ScalarQueryUtil with QueryCardinalityUtil
ScalarQueryUtil was wrong. Its goal semantic was to tell if given query
is scalar (always return a row), but it was returning true for queries
wrapped with FilterNode for which it is not known how many rows will be
returned.

QueryCardinalityUtil solves that as it returns a cardinality range,
where for scalar queries range is <1,1>.