Use a builder for WindowNode matching
The previous window matching pattern didn't validate all fields, and
didn't support a WindowFrame which contained Symbols (non-empty start or
end).  This commit borrows the builder idiom from PlanBuilder to create
flexible matching framework for WindowNode -- fields not mentioned to
the builder will not be validated.  Test code can simply supply the
constraints they wish to check.

This commit only moves the pre-existing constraints into a builder.  A
subsequent commit will add constraints for the rest of the fields of
WindowNode, including the aforementioned WindowFrame with symbols.