Introduce MultiSourceSymbolMapping
MultiSourceSymbolMapping contains list of source nodes, output symbols
and mapping of input symbols (the ones which come from source nodes) to
output symbols.

Motivation behind this is to use well defined entities instead of
complex collections.

The same thing exists in ExchangeNode, I will try to use
MultiSourceSymbolMapping there as a follow up of this.