Bound block sizes read from ORC files with metadata
Previously a block size is bounded by historical data collected when
loading a block. This approach can effectively bound 95% block sizes
according to the exported stats; however, It misses the first load that
could be as large as hundreds of MBs. This patch collects data in stripe
reader to estimiate the average row size in the next row group to read
and bound the block size with the estimation.