Make SpillContext subcontexts ownership explicit
When passing a `SpillContext` into a constructor, it's not obvious who
owns the object. In case of `SpillContext` this was caller's ownership
(`FileSingleStreamSpiller` would make a subcontext and use/close that)
whereas in case of `AggregatedMemoryContext` it was callee's ownership
(caller would make subcontext and pass that into the constructor).

This commit introduces a new pattern, where a supplier is passed into
the constructor clearing the ambiguity. The new pattern is applied to
`SpillContext` usage. Memory context usage changes will follow.