Prepare sources and make join node in the same method
It was not clear that one source of join node is prepared in one method
and that prepared sources was passed to the method which created another
source.

To make this more readable and clear, I joined that into a single method
which is now responsible for creating both sources and join node.