Add utility fail scalar functions
Sometimes scalar function that fails the query is useful.
Using conditional expression planner or even user may catch situations
that are not expected and can be only checked during the runtime.

Optimizer may rewrite the query into a form which is not exactly equivalent
and ensure that rewritten form is correct with check state conditions
and such fail functions. Notice a case of scalar subquery which is expected
to return single value, but it may return many values.