Fix humongous allocation when computing retained size for SliceArrayBlock
When computing the retained size of a SliceArrayBlock,
an IdentityHashMap will be created with expected size set to
the number of slices. For a SliceArrayBlock contains 1 million slices,
the IdentityHashMap will contain an Object array with 4,194,304 entires
which will cause humongous allocation.

Most of the slices will share the same base byte array, and setting
such huge expected size is a waste and cause GC issues. This commit
changes to use the default expected size.  This aligns with other usage
of IdentityHashMap including SliceArrayBlock.deepCopyAndCompact.