Introduce yield signals to operators
Splits can run for hours even after they have been killed. This is due
to expensive functions (e.g., regexp) or cross join. Generally, an
operator is not going to stop until it hits the page limit. The stop
condition is not subjective to time constraints given time checking is
at driver level. This patch adds a yield signal that will be set when
the time is up. An operator is actively checking the yield signal and
can stop even when the page is not full.
The patch adds yield signals to lookup join operators and cursor
processors