Implement HasSource { PlanNode getSource() } in rules having the method
This will allow for more type-safe matching. Right now it's possible
to write

    join().with(source().matching(...))

which should not be possible, as join has no single source.

Same goes for tableScan(), values(), exchange(), union() and others.

This common interface will allow defining a property specific to the
HasSource type, and rename the now-used unsafe property to onlySource()