Cap aggregation row count at input row count
The row count for aggregations is estimated as the product of the ndvs of each
grouping column. However, the number of output rows can't be greater
than the number of input rows.  When the product of the number of ndvs
is greater than the the input row count, estimate the output row count
as equal to the input row count.