Remove trailing spaces from CHAR function results
CHAR values are stored in Slices without the trailing spaces (and there is a checkState
for that). Some CHAR functions can produce values with a space at the end (for example,
substr if there are spaces in the CHAR that you are taking a substring of), so those
functions need to have the trailing spaces trimmed too.