Adding limited support for case-sensitive table names
Based on needs in issue https://github.com/prestodb/presto/issues/2863,
where in RDBMS that supports case-sensitive table names and have
tables with case-sensitive names, then those tables are unable to be
used with PrestoDB. This change adds optional (configuration controlled)
mapping for table names when executing queries that end up calling
getTableHandle(). It maps a lowercased name to the name of the table
from the database, and supports reloading the mapping in the event that
a table is referenced but doesn't currently have a mapping so new tables
can be used without restarting Presto.

A test was attempted to be written similar to the TestJdbcClient with a
new TestingDatabase, however it was discovered that H2 will return TRUE
for metadata.storesUpperCaseIdentifiers() even if
DATABASE_TO_UPPER=FALSE is set. The only way to have that return FALSE
is using MODE=MYSQL, however in that case H2 lowercases all table names.
So it is not possible to create a test with H2 without modifying the
section of code with the call to metadata.storesUpperCaseIdentifiers().