Extend PickTableLayout to pick layouts before AddExchanges
Layouts don't get chosen until AddExchanges. However, in order to have
better stats estimates, we need to know the table layouts earlier in the
plan.  We extend PickTableLayout to support picking layouts before
AddExchanges.  These layouts may not be the ones that are ultimately
chosen by AddExchanges, but we just need a best guess.