Support PredicatePushdown with distinct marker used in predicate
Typically MarkDistinct is used to provide distinct source
values (let say column A) to aggregate function. In order push down
predicate on column A over MarkDistinct you would need to push down it
over aggregation first, so that would mean that column A would must to
survive the aggregation which is not possible.

However, they might be optimizer rules which could use MarkDistinct
without aggregate function, in such case column `A` survives and it
accessible to futher nodes. This commit add support for this case in
PredicatePushdown.