Improve error message for functions involving TypedSet
This changes the exception type + message thrown when TypedSet.add()
exceeds the 4MB per-function memory limit.

TypedSet.add() used to throw an ExceededMemoryLimitException with an
error message about local memory limit exceeded, which was confusing
for users. The new exception message and code relate to this specific
case of a function exceeding memory usage.