Verify window frame using planner instead of parser object
For queries that have window frames embedded inside the
window definition of a named window, the Optional<Window>
object inside FunctionCall object will not carry the frame
information. Instead, the frame information is contained
in the planner object of type WindowNode.Frame inside the
window function.

This planner object carries the window frame information
in case of named windows as well as inlined windows. Hence,
we should use the planner object to verify window frame
information in the SQL query plan in the tests.

* Get rid of the Optional<WindowFrame> field from
ExpectedValueProvider<FunctionCall>, since it is only used
in tests.
* Change all the window function tests to use
WindowNode.Frame rather than WindowFrame