Optimize query planning for simple IN clauses
For query optimization, IN clauses are translated into nested OR clauses
and then processed in the normal OR clause path. For simple IN clauses
where it's just on simple literals though, this adds a lot of
unnecessary processing. This processing is also repeated multiple times
per query. By detecting the simple IN clause case the time processing
can be reduced significantly by short circuiting all that logic. Testing
a query with ~1 million IDs, planning took ~10 minutes. After this
change, planning is reduced to ~30 seconds.