Clean up PositionsPageProcessorIterator::processBatch() return value
processBatch() is responsible for telling the caller if it has produced
a page, yields to time, or needs to shrink the batch size. Currently
there is a local variable and a global variable to pass the information
back to the caller, which is not elegant. Unify the interface to return
a class to informm the caller the state returned by processBatch().