Rename NodePartitioning to TablePartitioning
Also rename ConnectorNodePartitioning to ConnectorTablePartitioning.

The concept of node partitioning will continue to exist in the context of
optimization to describe guarantees provided by remote exchanges and other
plan nodes. The commit removes the concept of node partitioning from table
layout.

TablePartitioning describe partitioning between split groups of a table.
The previous node partitioning describe partitioning between nodes.
The former implies the latter when execution of all splits in a split group
is guaranteed to happens on the same node, which was the case with node
partitioning and will continue to the case for the time being.

The new semantics immediately enables expression of partitioning within
a node when there are more split groups than nodes.

In the future, by decoupling split group partitioning from node partitioning,
it would be possible to remove the performance penalty incurred today
as a result of the now-legacy artificial requirement that each split group
must be scheduled on the same node.