Add inverse_normal_cdf function
This math function will help when analyzing datasets that
require the use of normal distributions.

Compute the inverse of the Normal cdf with given mean and
standard deviation for the cumulative probability: P(N < n).
The mean must be a real value and the standard deviation must
be a real and positive value. The probability p must lie on
the interval (0, 1).