Yield HashAggregationOperator if memory is not enough for rehash
Rehash can cause a steep 2X increase in memory usage. It is dangerous to
increase memory usage significantly before reserving it in memory
accounting system. This can lead to JVM OOM. We observed frequent full GCs
in production, which can be attributed to this root cause.  This patch
buffers unfinished page in the operator if the operator needs to rehash
the hash table when there is not enough memory. The operator will yield
and become blocked for memory allocation, and continue rehashing once
unblocked.