Make PartitionedLookupSource to close underlying LookupSources
Since `PartitionedLookupSource` encapsulates an array of `LookupSource`
instances, the `PartitionedLookupSource.close` should close all the
partitions too.

Currently the `LookupSource` implementations that have no-op `close` are
`OuterLookupSource` (used for RIGHT OUTER JOIN when there would be only
a single partition) and `IndexLookupSource` (used instead of
`PartitionedLookupSource` when JOIN is optimized as index-join), so this
commit effectively does not change anything, but makes code more
future-proof.