Expire specialized function cache entries after one hour
We have observed repeated compilation of MethodHandle that leads to
full GCs.  We notice that flushing the specialized function cache
mitigate the problem.

We suspect that it is a JVM bug that is related to stale/corrupted
profiling data associated with generated classes and/or
dynamically-created MethodHandles.

This might also mitigate problems like deoptimization storm or
unintended interpreted execution.