Add eager compact option into PagesIndex
Currently PagesIndex compacts the blocks lazily. It is only compacted
when the HashBuilderOperator or OrderByOperator hold it cannot reserve
the memory required.

However, compacting blocks requires to copy them. Thus when the cluster
is running low on memory and multiple operators is requesting memory,
all these operators will start to compact the PagesIndex and put a huge
pressure on the memory, which frequently causes full GC.

Adding the eager compact option into PagesIndex allows PagesIndex to
compact immediately whenever a new page is added.