Use dictionary mask to avoid block copy in join
A chain of joins in a single driver can lead to unnecessary copies of
blocks. The patch introduces masks to dictionary blocks so that during
the join phase, we generate dictionary ids and mask them on top of the
original block to be a dictionary view. For now, only the probe side is
with masks.