Fix column listing for tables with only unsupported columns
When a table accessed over a JDBC connector has no supported columns
(all columns are of some unsupported data type), we cannot allow users
to query such tables.  However, this should not prevent users from
querying `information_schema.columns`.

This commit fixes querying of the `information_schema.columns` when
there is a table with no supported columns, by removing the check in
`BaseJdbcClient`.  Since this check was needed to ensure we do not query
tables that appear to have no columns, equivalent check is provided in
`StatementAnalyzer`.  This is tested using `postgresql` connector, since
connectors usually used for tests do not support tables that appear to
have no columns.

Fixes #9434