Enable tests for non-legacy timestamp semantics
This enables non-legacy timestamp tests.

Remark:

At this point the tests are still failing.
The final commit which fixes the tests is
`Introduce new TIMESTAMP semantics to to_iso8601 scalar`

We cannot postpone enabling tests till after that commit. That is
because ignoring is done via commenting out `extends TestDateTimeFunctionsBase`
in test definition. And with it subsequent commits cannot modify the
non-legacy test classes keeping the code comilable.