Fix a race condition in Driver
If OperatorContext::requestMemoryRevoking() is called before the driver
thread calls updateDriverBlockedFuture(), the driverBlockedFuture will be completed
but right after that the driver thread will update it with a new future and
return it. However, the returned future is not completed even though memory revocation
is requested (we are missing a notification).