Use appropriate initial state object in generated aggregation code
The generated code made the base state for aggregations a
SingleState type. Combine for the aggregation had to handle merging
that empty single into the first group.

This makes it a group state for consistency and to simplify group
state's so they don't have to handle single state (shouldn't know
anything about them)