Add a priority column to the db resource group selectors
When using the json resource group configuration the selectors list
has a well-defined order (the order in which selectors are defined
in the config file). The order in which selector rules are checked
matters if we want a selection rule to have a higher priority than
the other rules. This change adds support for defining an order
among selectors for the db-based resource groups.