Finish ArbitraryOutputBuffer when all pages are read
The buffer only produces a single copy of each page for an arbitrary
reader, so we don't need to wait for no-more-buffers to finish. This
allows the stage to finish while the downstream stage is still scheduling
and thus has not set that flag, which is needed for writer scaling.