Split ClientBuffer::getPage into two interfaces
ClientBuffer used to have a single getPages() interface to return all
buffered pages to the caller. This approach is hard for the source to
decide which sinks to pull data from given the number/size of pages are
unknown beforehand. This patch separate the interface into getSummary and
getData so that the source may get a list of page sizes and acknowledge
the acceptance of pages through getSummary() and then fetch pages through
getData().