Pass tempto configuration as env variables
So far it was expected that tempto configuration files follow the files
naming convention. Base configuration file was called
tempto-configuration.yaml and profile was called
tempto-configuration-local.yaml.

Such approach is not extensible, it is hard to add the third
configuration layer (or more).

Instead of above approach, here the new convention is introduced.
Location of main configuration file is fixed, and location of profile
configuration file is provided with TEMPTO_PROFILE_CONFIG_FILE env
variable. For future layer new env varialbes will be introduced.