Filter out the columns that are partition columns from stats request
Filter out the partitioning columns from the column list that is sent in
the partition stats request to Hive Metastore, because those columns are
not treated as a regular column and are not listed on the column list,
causing some assertions to fail in the JDO code paths in Hive Metastore
(as opposed to the direct SQL code paths). See Hive Metastore's
`ObjectStore.getPartitionColumnStatisticsInternal` for details.