Fix bug in geometry_to_bing_tiles
The geometry_to_bing_tiles function when applied to a tile's polygon should return
a single tile, the input tile itself. However, this function was returning
four tiles: the input tile, a tile on the right, a tile at the bottom,
and a tile touching the lower right corner of the input tile. In general, it
could return more tiles than necessary when lower or right side of the bounding box
of the geometry was aligned with the tile border. This change fixes that.