Execute PickTableLayout before AddExchanges
Previously, the `PickTableLayout` rule was responsible for picking
layout in local test execution only.  However, in order to have
better stats estimates, we need to know the table layouts earlier
during the planning. We change PickTableLayout so that it
can pick (and re-pick) table layout iteratively.
Even though the rule follows `AddExchanges`'s logic, the intention
is not to overwrite layout chosen by `AddExchanges`.
Therefore now the rule should not be run after `AddExchanges`.