Improve split assignment performance using a custom future
One of the responsibilities of Presto coordinator is assigning splits to tasks.
The split assignment implementation involves cancelling a lot of futures.
Cancelling future is expensive because it involves filling in stack traces.
It can take more than 10% of overall CPU on a coordinator, causing
a performance bottleneck.

7a7bb28d attempted to fix this problem.
Unfortunately, it only solved half of the problem, and didn't
resolve the performance bottleneck.

The cancellation cause no longer has a stacktrace with this fix.
However, when invoked, `AbstractFuture.getDoneValue` would wrap
the cause with either `CancellationException` or `ExecutionException`,
and populates stacktrace on the wrapper exception.
For any useful `Future`, someone is directly or indirectly calling
`get` on it (even with `Futures.addCallback` for example).