Use whenAnyCompleteCancelOthers in Airlift
Split assignment on coordinator had performance issues.
The root cause is bad performance in getFirstCompleteAndCancelOthers.
Latest airlift fixed CPU waste, and introduced whenAnyCompleteCancelOthers.

Implementation wise, the performance optimization was internally applied
to whenAnyComplete. As a result, the performance improvement was picked up
when upgrading airlift version. This commit doesn't actually change the
performance. It is merely a refactor.