Consider whole predicate when picking layout
This fixes a performance regression due to a recent change to make
PickTableLayout run before AddExchanges (e6d963b1b32d3c5388a8f92ed4f04153b5570983).

For queries containing complex expression over partition keys, PickTableLayout was
not using this information when asking the connector for a layout. If a Hive
table has many partitions (several thousand), this causes the TableScan node to
contain a "current constraint" that enumerates all the possible values of the key.

Later, when AddExchanges.planTableScan executes, it builds an expression out of the
current constraint and any additional filters, which the Hive connector uses to prune
uninteresting partitions. Due to historical reasons, evaluating that expression can be
very expensive. It requires analyzing the expression, doing type inference, finding
function implementations, etc.

This change adds the ability for PickTableLayout to evaluate the full constraint in the
same manner AddExchanges used to do. Assuming the number of candidate partitions is small,
the resulting "current constraint" is smaller and cheaper to evaluate when AddExchanges
runs.