Lazy load buffer for large ORC streams
An 8 MB buffer will be created for a large ORC stream when reading a
new stripe. However the buffer may not be used given the data will not
necessarily be loaded. In production we found frequent full GC due to
fast allocation and release of such buffers. This patch delays the
buffer creation to when the data is to be loaded.