Apply Constraint.predicate on TPCH columns
Apply Constraint.predicate on TPCH columns which support predicate
pushdown.

The aim is to provide coverage for AddExchanges and other planner
components which leverage table layout selection and partition pruning.
It is expected that this change will make it easier to avoid or find
regression in planner related to table layout.

See:
 - com.facebook.presto.sql.planner.BenchmarkPlanner
 - com....optimizations.AddExchanges.LayoutConstraintEvaluator