Fix TPCH predicate pushdown when filtering column is pruned
Take a look at query `SELECT 1 FROM orders WHERE orderstatus = 'O'`.
In that case FilterNode will be optimized to `currentConstraint` in
`TableScanNode`, `TableScanNode` do not output any column.

In order to support that in TPCH we need to filter out all rows which do
not match `currentConstraint`. It is needed to have a
`TpchTable` being accessible in `TpchRecordSet` so we could check
column used in `currentConstraint` have values contained in the domain
of `currentConstraint`.