Measure communications timeouts from first failure
Backoff fails queries early because it measures duration from the
end of the last successful request.  In cases where there is a large
gap between request, a query may fail after a single attempt.
Instead the failure duration is measured from the end of the first
failed request in a sequence of failures.  With this change the
silding failure window is no longer possible, so all of the code
related to it has been removed.