Fix NULLIF for map type when type coercion is required
The compilation of NULLIF will perform necessary cast into common type.
It uses "dup" as the expression to generate the first value when invoke
the cast function, assuming the first value is on the stack. This
doesn't work when the invocation framework need to push other stuff
(e.g. ConnectorSession) before invoke the cast function.