Use HashTreePMap in PlanNodeStatsEstimate to reduce map copying
When stats calculator was actually put to use,  there was a lot of map
copying (through `ImmutableMap.copyOf(...)` and `new HashMap(...)`)
which was significantly impacting stats code performance.
`HashTreePMap` is much better for cases where individual entries of base
map are modified which is common case in stats code.

Following are results of the `BenchmarkReorderJoinsConnectedGraph`
benchmark class, which actually will be added later. The results are
based on CBO epic branch, which has more stats calculating rules, the
`ReorderJoins` iterative optimizer rule (i.e. a consumer of stats) and
said benchmark.

Results for `BenchmarkReorderJoinsConnectedGraph.benchmarkReorderJoins`
on 8 tables with COST_BASED join reorder strategy:

- with `ImmutableMap`s:

    Benchmark                                                  Mode  Cnt    Score    Error  Units
    BenchmarkReorderJoinsConnectedGraph.benchmarkReorderJoins  avgt   30  579.970 ± 15.814  ms/op

- with p-collections:

    Benchmark                                                  Mode  Cnt     Score    Error  Units
    BenchmarkReorderJoinsConnectedGraph.benchmarkReorderJoins  avgt   30  1540.031 ± 40.562  ms/op