Improve memory tracking of PartitionedOutputOperator
If the memory usage is updated only in the constructor and not in
addInput the usage can be off by several orders of magnitude. However,
we cannot update the memory usage at every addInput call by calculating
the retained size of the PageBuilders in PagePartitioner as that can be
expensive especially for complex types. Therefore, this change
approximates the memory usage by using getSizeInBytes() instead of
the more expensive getRetainedSizeInBytes() method.

With this change there is almost no difference in performance:

Benchmark                       (pageCount)  Mode  Cnt     Score     Error  Units
updateMemoryUsageInAddInput            5000  avgt   40  2962.434 ± 151.650  ms/op
updateMemoryUsageInConstructor         5000  avgt   40  2896.495 ± 217.911  ms/op