Fix race-condition when Driver re-consults Operator.isBlocked
Fixes second problem described in
https://github.com/prestodb/presto/issues/10003:

Previously, if `Driver` couldn't make progress (`!movedPage`), we listed
all blocked operators and waited until any of these unblocked. However,
IF an operator become unblocked between the time when it was considered by
the processing loop AND the time when we construct driver's composite
is-blocked future THEN the this fact is lost. Thus, even though an
operator became unblocked, driver will still be blocked, waiting for
something else.