Fix race in OutputBufferMemoryManager::updateMemoryUsage()
It's important that setBytes(bufferedBytes.addAndGet(bytesAdded)) is
synchronized as it's a compund action. Otherwise, when pages are
dereferenced we can end up with incorrect accounting.

Here is an example. Assume that bufferedBytes is 100 bytes and
two pages are dereferenced, say with bytesAdded = -5 and -10, so the
expected reserved memory in the end is 100 - (5 + 10) = 85 bytes.
Assume that thread T1 runs bufferedBytes.addAndGet(-5) and gets 95,
then T2 runs bufferedBytes.addAndGet(-10) and gets 85
(as bufferedBytes is atomic). Immediately after that T2 can execute
setBytes(85) and then T1 can execute setBytes(95) after T1, and we end
up with 95 bytes in the end, which is incorrect.