Remove Inner Classes
The original memory improvements used inner classes to add abstractions
to make it more clear we were working with a hash table. While, these
added negligible GC pressure as measured with JFR, this is a
dangerous anti-pattern for code in tight loops, so this commit
removes the objects and makes method calls instead.

NOTE: when measured with JMH, in a synth benchmark, a 600 MB/s
alloc rate (which still had no impact on presto) goes almost to 0.