Add named variable support to resource group templates
This is accomplished by changing the ResourceGroupConfigurationManager
interface in the SPI to return a SelectionContext containing the
resource group and a generic type. This allows the engine to pass in
the SelectionContext to the configure method later, where the original
context can be re-used by the plugin to do the right thing.

The built in resource group managers use a VariableMap to augment the
context.

Administrators can now define their own custom variables (by using
Java named capturing groups in their source/user regex), and use these
variables in the selector resource group ID and in the resource group
tree.