Fix missing unblock signal in SourcePartitionedScheduler
Among the things that SourcePartitionedScheduler.schedule returns is
a ListenableFuture. The future should be completed whenever calling schedule
again may lead to material action (e.g. more splits being scheduled,
or notification that no more split will be produced).
A bug was found that SourcePartitionedScheduler sometimes does not complete
the future when no more splits will be produced. This commit fixes that.

Note that the user visible behavior change of this commit is very limited
due to the 1 second timeout failsafe in SqlQueryScheduler.