Minor Cleanup of Histogram and HistogramState
Histogram had dead code in that state.get() could not return
null any more, so the if/else checks in Histogram.combine() were
superfluous and removed. I did leave in assertions
that state.get() != null as the current code is not written to
handle such a case.

Change also removes the Histogram.set() method entirely as
it is not used any longer.

Lastly, In order to make sure I was getting consistent testing,
I made TestHistogram use a single code path of functions to get
it's InternalAggregationFunctions.