Add maven flag to create tests jar with dependencies
Sometimes it is usefull to be able to run a test or a benchmark on the other
machine than a development one. For example it is prefered to run benchmarks on
a solid desktop or on a server machine to overcome overheating deviations in benchmark
execution time.

Since most of the modules have complex dependency tree it is very hard to copy
every dependency to the remote server and write a cmd line that has all the dependencies
included.

This commit is supposed to address this issue.

Now we can tell maven to assemble an uber(fat) jar that includes the entire "test" scope
classpath for a particular module.

For example after running:

./mvnw clean install -P tests-with-dependencies -pl presto-geospatial

The uber jar that contains the entire class path of the "test" scope of the "presto-geospatial"
module is created and placed to that modules target directory:

./presto-geospatial/target/presto-geospatial-0.197-SNAPSHOT-tests-with-dependencies.jar

Now using this jar we can easly run any benchmark from that module via the command line:

java \
  -cp ./presto-geospatial/target/presto-geospatial-*-tests-with-dependencies.jar \
  com.facebook.presto.plugin.geospatial.BenchmarkSTIntersects