Pick Benchmark* files as potential TestNG tests
It happens that data used for benchmarking is not trivial and unit testing
of the data is sometimes needed.

It is better to keep both, data and the test in the same place - a benchmark
file.

By default SurefirePlugin (a maven plugin that runs TestNG as part of a build)
picks only files that start with "Test" prefix and few other variations that
include a "Test" word: https://goo.gl/KSjvVb

By introducing that commit we are making SureFire to pick classes with a
Becnhmark* prefix as well, so TestNG could run methods marked as @Test