Optimize DECIMAL constants for byte code generation
Since `LiteralInterpreter.toExpression`
no longer represents the `DECIMAL` using "magic literal function", the
byte code generated for a `DECIMAL` constant was not optimal.  This was
because `ExpressionOptimizer` knew how to fold constant functions, but
did not know how to fold constant CASTs.

This commits improves `ExpressionOptimizer` abilities, so that `DECIMAL`
constants are represented in the byte code in simpler way.