Move null replication check from PropertyDerivation to AddExchanges
This check validates plan correctness. However, property deriviations
are not the best place for enforcing correctness (they should just
compute properties). It should be rules or validators that check
correctness. This commit moves null replication check to AddExchanges
so that the condition can be verified on the go as the exchanges
are added.