Modifies TypeSignature behavior for ROW types with quoted identifiers as field names.
The TypeSignature for ROW fields currently includes quotes if the
identifier was originally quoted. This is problematic if a single
query contains quoted and unquoted versions of the same field.
For example the query:
    select cast(row(11) as row(field1 bigint)) as col1
        union select cast(row(12) as row("field1" bigint));

Results in the output field 'field0'
        col1
    -------------
     {field0=12}
     {field0=11}

If any reserved words are used as field names then quoting them becomes necessary.
This change modifies TypeSignature construction to ignore any quotes
around a field.