User lower case when listing schemas
SchemaTableName or QualifiedObjectName either make sure that name parts
are in lower case or make them to be in lower case. However when listing
schemas, schema is kept in plain String which does not enforce name to
be in lower case.

To make schema consistent with other name related objects it is
transformed to lower case in MetadataManager. That way it will work for
all connectors.

Alternatively, a SchemaName could be introduced. But this is much bigger
change and may need SPI to be changed as well.