Support PredicatePushdown with distinct marker used in predicate
Typically MarkDistinct is used to provide distinct source
values, say column A, to aggregate functions. In order to push down
predicate on column A over MarkDistinct we need to push it down
over the aggregation first, so that means column A must survive the
aggregation which is not possible.

However, there may be optimizer rules that use MarkDistinct
without aggregate functions, in those cases column A survives and it
is accessible to other nodes. This commit adds support for this case in
PredicatePushdown.