Do not derive all grouping sets when traversing tree
The only node that has visitable expressions is SimpleGroupBy. There's
no point in synthesizing the grouping sets from nodes like Cube or Rollup
just for the purpose of visiting the resulting expressions.

One effect of doing so is that query creation can fail due to OOMs
when allocating too many objects.