Remove documentation for log function
The function was removed in b3fd8b1e951c2f7c4aa55b19b152b670781fde5d.
The arguments to the function were in the wrong order according to the
SQL standard, resulting in incorrect results when queries were
translated to or from other SQL implementations. The equivalent to
`log(x, b)` is `ln(x) / ln(b)`. The function can be restored with
the deprecated.legacy-log-function config option.