Fix ST_IsValid and geometry_invalid_reason functions
Re-implement ST_IsValid and geometry_invalid_reason functions using ESRI
library.

The original implementation was failing with "shell is null but hole found"
error for some self-intersecting geometries when deserializing geometry Slice
into JTS Geometry object. A polygon consists of an exterior shell and zero or
more interior holes. A multi-polygon can have multiple shells, each with zero
or more holes. Slice representation of a geometry contains a list of exterior
and interior rings, but doesn't identify exterior rings explicitly. It is only
known that points of exterior rings are listed in clockwise order, while
points of interior rings in counterclockwise order. In some cases, e.g.
hourglass shaped polygon POLYGON ((0 0, 1 1, 0 1, 1 0, 0 0)), it is not
possible to determine whether a set of points is listed in clockwise or
counterclockwise order. In these cases, deserialization to JTS geometry object
fails.

This commit updates ST_IsValid and geometry_invalid_reason functions to use the
functionality from the ESRI library and avoids deserialization to JTS Geometry
object.