Fix various broken web links
Due to changes in cd3b2a8 some existing links (e.g., /tableau)
and the query uri generated by the presto cli was broken. This
change fixes the first issue by introducing a WebUiResource
and doing a redirect, and fixes the second issue by prepending
the generated uri with '/ui'.