Fix column access checks for views
Separate the column access checks we need by identity and only check
privileges for the current session identity.  This ensures that we keep
the checks for the view owner (who needs permissions on all the tables
that make up the view) separate from the checks for the user querying
the view (who only needs permissions on the view itself).