Push up AssignUniqueId through remote exchange
Scalar subqueries are often rewritten as follows:

- Aggregation(unique, a, b)
  - LeftJoin
    - RemoteExchange (REPARTITION)
      - AssignUniqueId(unique)
        - probe (a, b)
    - RemoteExchange (REPARTITION)
      - build

In these plans, streaming aggregation is not applied because
grouped(unique, a, b) local property generated by the AssignUniqueId node
can't reach the Aggregation node as it gets dropped when passing through the
RemoteExchange node.

This commit adds an optimizer rule to push up AssignUniqueId node through the
RemoteExchange to allow for streaming aggregation. It also updates
HashGenerationOptimizer to avoid generating unnecessary hashes for streaming
aggregations. The modified plan looks like this:

- Aggregation(unique, a, b)
  - LeftJoin
    - AssignUniqueId(unique)
      - RemoteExchange (REPARTITION)
        - probe (a, b)
    - RemoteExchange (REPARTITION)
      - build