Allow at most one public constructor in @ScalarFunction class
A @ScalarFunction class declares one SQL function. The class
may have multiple methods, allowing generic specialization.
In the @ScalarFunction class, even though the SQL function signature
is declared once in each method, they are required to be compatible.

As a result, it doesn't make much sense to allow multiple constructors,
each taking a different set of type parameters. Regardless of which method
is chosen, the bound variables map is the same.

If multiple constructor is found to be useful going forward,
a mechanism could be introduced allowing each method to specify their
corresponding constructor. This likely won't be necessary because
the function implementor can branch his/her code based on the bounded
variables. The current mechanism requires that such implementors have
two methods with different set of TypeParameters, which generally
cannot be satisfied.