Limit the size of local buffers for stream readers
In the past we have seen large number of humongous allocations on the
cluster worker nodes. The size of each allocated buffer could be as high
as 80MB, and the total allocation could be as high as 100GB per worker
node. This situation happens when reading map or array types. The reason
of such high memory consumption is two folds:

1) When we read the nested types like map and array, we didn't take the
number of elements within each cell into account when caculating the
batch size. A single map cell could contain 20K entries in some cases.
Currently we read the first block and bound the batch size accordingly,
but the first block might already contain too many (e.g. millions of)
entries.

2) The number of stream readers Presto creates for each task involving
scan could be a lot. It is the product of the following numbers:

    a) The number of drivers that was assigned with a split to read.
    b) The number of columns that need to be scanned.
    c) The number of streams for each column. This depend on the column
       type, e.g. the MapStreamReader contains keyStreamReader and
       valueStreamReader.

    This could be up to hundreds stream readers. Each stream reader hold
a few of local buffers depending on the underlying streams. These local
buffers were of the same size of the batch size, so if we mis-estimated
the batch size to be some large number, these local buffers would be
humongous too.

The fix is 3 steps:

1) In each stream reader, we limit the size of the local buffer to 1K
rows at maximum no matter how large the batch size is. The blocks
being built would still be humongous but we can reduce the local buffer
sizes. We also removed the lengthVector and reuse the offsetVector
to read in the length values, then convert them to the offset values
whenever possible.

2) Read the stats for map and array columns, and use it as a heuristic
to determine the batch size for stream readers. Currently we only did
this for string type but not map and array types.

3) If the stats is not available, we could try to start from very few
rows to determine how many rows to read for the first block.

This commit is only for step 1).