Improve Stripe Size Estimation for SliceDictionaryColumnWriter
When writing dictionary encoded columns, the ORC write estimates
the buffered size based on (1) estimated index bytes per entry,
and (2) number of entries.

However, this doesn't account for the compression, which can be
very efficient when the dictionary indices are highly skewed.
For an extreme example, thinking about dictionary has 1M entires,
yet one value appeared all the time, while all the other (1M - 1)
entires only appeared once.

We have observed written stripe size is only 10M ~ 20M when the
estimated stripe size is full at 64MB for real workloads..

This commit makes SliceDictionaryColumnWriter also writes the
dictionary index to a temporary data stream, and use the buffered
size of that stream to estimate stripe size.

Note dictionary indices are now written in two places: the IntBigArray
(values) and the temporary data stream. This is because dictionary
has to be sorted and dictionary ids has to be re-indexed.

A potential future work to reduce memory is to read back
the dictionary index from the temporary data stream.